src/table-updater/requirements.txt
pandas==1.5.1
pyscbwrapper==0.1.1
requests==2.28.1
sqlalchemy==1.4.42
google-cloud-pubsub==2.13.10
psycopg2==2.9.4
